+++ packages/lumx-core/src/js/types/index.ts
@@ -25,4 +25,5 @@ export type { HasChecked } from './HasChecked';
 export type { HasDisabled } from './HasDisabled';
 export type { AriaAttributes } from './AriaAttributes';
 export type { Selector } from './Selector';
+export type { PropsToOverride } from './jsx/PropsToOverride';
 export type { PartialBy, PartialExcept } from './PartialBy';