+++ packages/lumx-core/src/js/types/jsx/PropsToOverride.ts
@@ -1,0 +1,2 @@
+/** list of generic props defined on JSX that need to be redefined for each framework */
+export type PropsToOverride = 'ref' | 'onClick' | 'onChange' | 'onKeyPress';