+++ packages/lumx-react/src/components/checkbox/Checkbox.tsx
@@ -14,11 +14,15 @@ import { useMergeRefs } from '@lumx/react/utils/react/mergeRefs';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 import { useDisableStateProps } from '@lumx/react/utils/disabled/useDisableStateProps';
+import { ReactToJSX } from '@lumx/react/utils/type/ReactToJSX';
 
 /**
  * Defines the props of the component.
  */
-export interface CheckboxProps extends GenericProps, Omit<UIProps, 'inputId'> {}
+export interface CheckboxProps extends GenericProps, ReactToJSX<UIProps, 'inputId'> {
+    /** On change callback. */
+    onChange?(isChecked: boolean, value?: string, name?: string, event?: React.ChangeEvent): void;
+}
 
 /**
  * Component default props.