+++ packages/lumx-react/src/components/icon/Icon.tsx
@@ -3,10 +3,11 @@ import { GenericProps } from '@lumx/core/js/types';
 
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 import { useTheme } from '@lumx/react/utils/theme/ThemeContext';
+import { ReactToJSX } from '@lumx/react/utils/type/ReactToJSX';
 
 export type { IconSizes };
 
-export interface IconProps extends Omit<UIProps, 'children'>, GenericProps {}
+export interface IconProps extends ReactToJSX<UIProps>, GenericProps {}
 
 /**
  * Icon component.