+++ packages/lumx-react/src/components/message/Message.tsx
@@ -1,11 +1,9 @@
 import { forwardRef } from '@lumx/react/utils/react/forwardRef';
 import { Message as UI, CLASSNAME, COMPONENT_NAME, MessageProps as UIProps } from '@lumx/core/js/components/Message';
 import { GenericProps } from '@lumx/core/js/types';
+import { ReactToJSX } from '@lumx/react/utils/type/ReactToJSX';
 
-export interface MessageProps extends GenericProps, Omit<UIProps, 'children' | 'ref'> {
-    /** Content. */
-    children?: React.ReactNode;
-}
+export interface MessageProps extends GenericProps, ReactToJSX<UIProps> {}
 
 /**
  * Message component.