+++ CHANGELOG.md
@@ -33,6 +33,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     -   Uniformize documentation of `disabled` (as an alias to `isDisabled`) props across core components
 
 -   `@lumx/core`, `@lumx/react`, `@lumx/vue`: improve declared ESM exports in package.json
+-   `@lumx/core`, `@lumx/react`, `@lumx/vue`: rework types and interfaces inheritance
 
 ## [4.3.1][] - 2026-02-11